+++ ui/resources/webapp/view/Carrier.view.xml
@@ -10,52 +10,6 @@
 	<sf:SmartFilterBar id="sFb" entityType="FlightCarrierType" visible="true">
 	</sf:SmartFilterBar>
 	<st:SmartTable id="anaTab" smartFilterId="sFb" entitySet="FlightCarrier" useExportToExcel="true" tableType="AnalyticalTable" 
-					enableAutoBinding="true"
-					app:p13nDialogSettings="{
-												columns : { 
-													visible : true,
-													payload : { 
-														visibleItemsThreshold : 3,
-														items : [ {
-															columnKey : 'CARRID',
-															visible : false
-														} ]
-													}
-												}
-											}">
-		<!--<st:customData>
-			<core:CustomData
-				key="p13nDialogSettings"
-				value='\{ "group" : 
-							\{ 
-								"visible" : false
-							} 
-						}'>
-			</core:CustomData>
-		</st:customData>-->
-		<t:AnalyticalTable id="dashboardTable">
-			<t:columns>
-				<t:AnalyticalColumn id="carrid" managed="true" leadingProperty="CARRID" filterProperty="CARRID" sortProperty="CARRID" inResult="true"  width="150px" grouped="true">
-					<m:Label text="Carrier Id" />
-					<t:customData>
-						<core:CustomData key="p13nData" value='\{ "columnKey" : "CARRID", "visible" : "false" \}'></core:CustomData>
-					</t:customData>
-				</t:AnalyticalColumn>
-				<t:AnalyticalColumn id="flDate"  inResult="true" width="150px" leadingProperty="CARRNAME">
-					<m:Label text="Carrier Name" />
-					<t:customData>
-						<core:CustomData key="p13nData" value='\{ 
-																	"columnKey" : "CARRNAME" , 
-																	"columnIndex" : "2", 
-																	"sortProperty" : "CARRNAME"
-																\}'>
-						</core:CustomData>
-					</t:customData>
-					<t:template>
-                         <m:Text text="{ path : 'CARRNAME' }" />
-                	</t:template>
-				</t:AnalyticalColumn>
-			</t:columns>
-		</t:AnalyticalTable>
+					enableAutoBinding="true" >
 	</st:SmartTable>
 </mvc:View>
